requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
python-dotenv>=1.0.0
openai>=1.13.3
pydantic>=2.5.0
